Fix V3 interop compatibility and RLE decoder overflow issues
**V3 Metadata Reading**:
- Add DeserializeV3 methods for MetadataIndexNode and DeviceMetadataIndexEntry
- Support both Java V3 (PlainDeviceID) and C# simplified V3 formats
- Auto-detect V3 format by attempting Java footer parse first
- Extract device names from V3 index tree to create virtual schemas

**V3 Query Support**:
- Add QueryV3Simplified for C#-written V3 files (chunk scanning)
- Extend QueryV4 to handle V3 files using index navigation

+++ csharp/src/Apache.TsFile/Encoding/Encoder/Ts2DiffEncoder.cs
@@ -128,7 +128,7 @@ namespace Apache.TsFile.Encoding.Encoder
             }
 
             // Bit-pack deltas into encodingBlockBuffer
-            int encodingLength = (int)Math.Ceiling(packNum * packWidth / 8.0);
+            int encodingLength = (int)Math.Ceiling((long)packNum * packWidth / 8.0);
             byte[] encodingBlockBuffer = new byte[encodingLength];
             for (int i = 0; i < packNum; i++)
                 LongToBytes(_deltaBlockBuffer[i], encodingBlockBuffer, packWidth * i, packWidth);
